swml/parsing/tables-1.dat

#data
,
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>

#data
,==
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>

#data
,a,==
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             colspan="2"
|             "a"

#data
,==,==
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             colspan="2"

#data
,a,==,==
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             colspan="3"
|             "a"

#data
,==,==,a,==
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             colspan="2"
|           <td>
|             colspan="2"
|             "a"

#data
,"a\"b","c\\d"
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "a"b"
|           <td>
|             "c\d"

#data
,"a\"b\"cd\E\f","c\\\\d"
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "a"b"cdEf"
|           <td>
|             "c\\d"

#data
,*,a,*
,*"","*"
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <th>
|           <td>
|             "a"
|           <th>
|         <tr>
|           <th>
|           <td>
|             "*"

#data
,*b,* a,**a
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <th>
|             "b"
|           <th>
|             " a"
|           <th>
|             "*a"

#data
,*a*,*==
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <th>
|             "a*"
|           <th>
|             "=="

#data
,*ab,foo
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <th>
|             "ab"
|           <td>
|             "foo"

#data
,*a[SPAN[b,foo
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <th>
|             "a"
|             <span>
|               "b"
|           <td>
|             "foo"

#data
,*ab,==,foo
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <th>
|             colspan="2"
|             "ab"
|           <td>
|             "foo"

#data
,*ab,==,==,foo
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <th>
|             colspan="3"
|             "ab"
|           <td>
|             "foo"
